environment config added

+++ src/microsplit_reproducibility/datasets/custom_dataset_2D.py
@@ -63,13 +63,13 @@ def get_train_val_data(
     # val_idx = train_idx
     # test_idx = train_idx
     if datasplit_type == DataSplitType.All:
-        data = data.astype(np.float64)
+        data = data.astype(np.float32)
     elif datasplit_type == DataSplitType.Train:
-        data = data[train_idx].astype(np.float64)
+        data = data[train_idx].astype(np.float32)
     elif datasplit_type == DataSplitType.Val:
-        data = data[val_idx].astype(np.float64)
+        data = data[val_idx].astype(np.float32)
     elif datasplit_type == DataSplitType.Test:
-        data = data[test_idx].astype(np.float64)
+        data = data[test_idx].astype(np.float32)
     else:
         raise Exception("invalid datasplit")
